Add the get-interface-information operationnal call with terse param only
This call is helpful in listing interfaces without configurations in the datastore.

In order to behave like a real switch where stock ports (in a default state) does
not exist, the part that use to add ethernet switching everywhere has been removed.
It was thought to be useful for having empty interfaces show up even with default
values but that was the wrong assumption.  Ports with default values don't show up

The tests are also now profiting from the new delatble interface mecanism in the
cleanups

Both normal and qfx have been fixed an also had to fix the fact that the native-vlan-id
was not appearing if there was no ethernet-switching on qfx, which can happen